fix: use nr_price BIGINT in schema to match int64 price type

+++ migrations/001_initial_schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS tb_product (
     pk_product UUID PRIMARY KEY,
     tx_name TEXT NOT NULL,
     tx_description TEXT NOT NULL DEFAULT '',
-    nr_price DOUBLE PRECISION NOT NULL,
+    nr_price BIGINT NOT NULL,
     tx_image_url TEXT NOT NULL DEFAULT '',
     fk_category UUID NOT NULL REFERENCES tb_category(pk_category) ON DELETE CASCADE,
     ts_product_created_at TIMESTAMP NOT NULL,
